+++ spotinst_sdk/spotinst_functions.py
@@ -3,8 +3,10 @@ import json
 import os
 import tempfile
 import zipfile
+import logging
 
 none = "d3043820717d74d9a17694c176d39733"
+logger = logging.getLogger(__name__)
 
 
 # region Application
@@ -71,11 +73,20 @@ class FunctionCreationRequest:
         self.should_print_output = print_output
         self.function = self.rebuildFunctionInlineCode(function)
 
-    def print_output(self, output):
+    def print_output(self, output, level="debug"):
         if self.should_print_output is True:
-            print(output)
+            if level == "debug":
+                logger.debug(output)
+            if level == "info":
+                logger.info(output)
+            if level == "warn":
+                logger.warn(output)
+            if level == "error":
+                logger.error(output)
+            if level == "critical":
+                logger.critical(output)
 
-    def rebuildFunctionInlineCode(self, function):
+    def rebuildFunctionInlineCode(self, function, ):
         directory = function.directory
         handler = function.handler
 
@@ -91,20 +102,20 @@ class FunctionCreationRequest:
         return function
 
     def zip(self, src, dst):
+        self.print_output("info level", "info")
         zf = zipfile.ZipFile(dst, "w", zipfile.ZIP_DEFLATED)
         abs_src = os.path.abspath(src)
         for dirname, subdirs, files in os.walk(src):
             for filename in files:
                 absname = os.path.abspath(os.path.join(dirname, filename))
                 arcname = absname[len(abs_src) + 1:]
-                self.print_output(
-                    "collecting file {}".format(
+
+                self.print_output("collecting file {}".format(
                         os.path.join(
                             dirname, filename)))
                 zf.write(absname, arcname)
         zf.close()
 
     def toJSON(self):
-        del self.should_print_output
         return json.dumps(self, default=lambda o: o.__dict__,
                           sort_keys=True, indent=4)