+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.39] - Unreleased
+## [1.0.41] - UNRELEASED
+### Added
+- Logger using standard logging module
+- "--log-level" command line argument (more info in "--help")
+- "log_level" constructor parameter, defaults to "critical"
+- Created changelog
+
+## [1.0.40] - 2018-20-22
 ### Added
  - kubernetes api calls
  - b/g deployment api calls
@@ -35,3 +42,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - `resume_stateful_instance`
  - `pause_stateful_instance`
  - testing for `__init__.py`
+ 
+